Fix incorrect invocation when CLI is installed via pip3
When installed via pip3, the cli does not invoke the main entry point of the program via `__main__`.

Instead, it expects the following methods to be declared within `setup.py`:
* `conductr_cli.conduct:run`
* `conductr_cli.sandbox:run`
* `conductr_cli.shazar:run`

The change is to reinstate this behaviour correctly so the main files can be invoked from both `__main__` as well as pip3 installation.