Support the new Docker for Mac and Windows
The new Docker for Mac and Windows environments use a local hypervisor instead of docker-machine/boot2docker. As such we now check whether we can run docker at all prior to docker-machine/boot2docker. If we can then we need not check any further, and can assume docker's IP to be the loopback address.