Suppress stderr output from Docker commands where appropriate
The stderr output will be displayed if running CLI from a terminal where Docker env is not available.

The suppressed stderr ouput are from Docker commands which are used for detection, i.e. `docker info` for detecting host address to be set for CONDUCTR_IP. It is benign the suppress the stderr for these commands, since there are actions to be taken should the commands fail.

An example of this scenario is running CLI from a bastion host where Docker is not installed against a remote ConductR core. Prior this change, docker related error will be displayed on the terminal which is not productive in this case.