Cache directory migration
The PR https://github.com/typesafehub/conductr-cli/pull/161 creates now the default cache directory with the permissions 700 and validates if the specified cache directory has these secure permissions.

If a user has used a previous conductr-cli version than the default cache directory has been created with the permissions 755. This would lead to an error message when loading a bundle.

This PR is now checking if the specified cache directory during `conuct load` is the default cache directory, e.g. `~/.conductr/cache`. If that's the case and if the directory does not pass the secure permission check then we now set the permissions for the directory to 700.

Additionally, every file inside the cache directory is now created with the permissions 600 instead of 644.