Waitinf for conductr progress bar
Introducing a progress bar when running the `sandbox run` command and waiting that ConductR has been successfully started. The progress bar is the same as with the `conduct install` command.

In case a timeout occurs, now a dedicated error message is displayed.

Also, we are giving now ConductR 30 seconds time to start.