Fix the Docker Machine check
The existing check for Docker Machine was inaccurate. It is a valid scenario to have a native Docker installed, along with DOCKER_HOST and others e.g. in the future when connecting the Docker API with ConductR.

The only check required for Docker Machine is the DOCKER_MACHINE environment variable.