Fixed request dependency
The prior declaration of the request dependency was in two places - I don't know if this confused pip, but what I've now done is moved the min requirement for dcos to be our min requirement. This min requirement should also be less than what python3 on Debian insists and should thus be satisfied (v.2.9.1 is globally installed for python3 on Ubuntu).