Cleanup cached bundles after conduct load command
Cleanup will be based on the name and the compat version of the recently loaded bundle.

The following will be kept:
- The recently loaded bundle.
- The latest version of the bundle in the cache having the same name and compat version as the recently loaded bundle.

All other versions of the bundle having the same name and compat version will be removed.

The latest version is decided by the timestamp of the bundle file on the disk.