Use command prefix when starting features
The `conduct_main.run` has arguments that can be used to tell it which `conduct` sub command it should run, e.g. `conduct info`.

This function is called in 3 different ways and they all prexi different number of command before the actual `conduct` sub command:
1. Programtically, from another python function: No prefix argument, e.g. `conduct info`
2. When the user submits a `conduct` command: 1 prefix argument, e.g. `/Users/mj/workspace/dcos-cli/cli/env/bin/conduct conduct info`
3. When the user submits a `dcos conduct` command: 2 prefix arguments, e.g. `/Users/mj/.dcos/subcommands/conductr/env/bin/dcos-conduct conduct info`

Before parsing the command, all prefix arguments need to be removed. The PR is fixing this logic. Prior to this, the first element of the `_args` array has been also removed if the function was called programatically (case 1)