Disallow option combination --no-wait and --feature
When ConductR is started with features, the feature startup is depend on waiting that ConductR has been started. Otherwise, the feature startup will fail.

Therefore, this PR is adding a check that the options `--no-wait` and `--feature` are not set together. If that is the case, `sandbox run` will abort with an error message.