Stronger file check
This commit strengthens the check to determine whether a bundle exists on the file system by ensuring that it has a ".zip" extension. Prior to doing this the `load` sub command could yield a false-positive and cause the cache check to be avoided when a directory of the same name as the bundle existed. This problematic scenario would occur when using the load sub command from within the ConductR project itself given that there are directories for eslite, visualizer etc. CI would then also keep resolving bundles from bintray despite them possibly being cached from other tests.