Ensure address is available during sandbox run
When `sandbox run` is invoked for ConductR 2, perform checks to ensure ip addresses is available to each available instance given the address range.

The check is done by performing a socket bind to an address within the specified address range. If the bind is successful, the socket is immediately closed. Otherwise, the address is marked as unavailable.

If addresses can't be allocated to all core and agent instances, the `sandbox run` will fail and produce the commands which can be copy-pasted by the end user to setup the required address aliases.

The `--interface` parser is removed as well since the interface name is used only to print the user instructions in the scenario above and has no further functionality within the application.

The Docker env validation is now going to be performed for `sandbox run` ConductR 1 only.