Stop sandbox JVM mode
Implementation of the `sandbox stop` command for the ConductR JVM binaries.

The `sandbox stop` command is stopping both the docker containers and the JVM processes. If no containers or processes are running then no output is produced. If either the JVM or Docker stop function cannot stop the processes or containers respectively then an error is reported and the stop function returns False.

This PR also introduces additional unit tests for the `sandbox_stop_docker` module to validate the scenarios:
- No containers running
- Cannot stop containers