Docker validation for logging feature
In ConductR 2.x (JVM mode) the conductr-kibana bundle is a Docker based bundle. Therefore, when the logging feature is specified, it is checked if Docker is installed and correctly configured.

The logging feature uses the same Docker validation and error output as the sandbox run feature when using the Docker related ConductR binary.

An additional Docker validation for the monitoring feature is not necessary because it depends on the logging feature, i.e. the logging feature will be started first and will fail if Docker is not setup correctly.

**Docker validation output**

```
$ sandbox run 2.0.0-rc.2 --feature logging
|------------------------------------------------|
| Starting ConductR                              |
|------------------------------------------------|
Extracting ConductR core to /Users/mj/.conductr/images/core
Extracting ConductR agent to /Users/mj/.conductr/images/agent
Starting ConductR core instance 0 on 192.168.10.1..
Starting ConductR agent instance 0 on 192.168.10.1..
Waiting for ConductR to start.......
…
|------------------------------------------------|
| Starting logging feature based on              |
| elasticsearch and kibana                       |
|------------------------------------------------|
conductr-kibana bundle is packaged as a Docker image. Checking Docker requirements..
Error: Docker native is installed but not running.
Error: Please start Docker with one of the Docker flavors based on your OS:
Error:   Linux:   Docker service
Error:   MacOS:   Docker for Mac
Error:   Windows: Docker for Windows
Error: A successful Docker startup can be verified with: docker info
```