Set scheme to https if auth is enabled
If in the `settings.conf` file `conductr.auth.enabled = true` then we now set the `—scheme` mode to `https`. Using Basic Auth without https is not allowed. So this PR improves the user experience. The user does not specify `—scheme https` manually for each conduct command if auth is enabled.

Note that the user can not override the `—scheme` to http if auth is enabled. From an implementation point of view it is not possible to set a default of `—scheme` based if Basic auth is enabled because at the point in time we set the default we do not have access to any other arguments, e.g. if dcos is enabled or not which is also part of the logic. Therefore the default for `—scheme` is still HTTP and if Basic auth is enabled then we set `args.scheme` to `https`.

Fixes https://github.com/typesafehub/conductr-cli/issues/250.