Offline mode
** Issue description**
In the past the `conduct load` command always retrieved the version of a specified bundle name from Bintray. With the resolved version it checked if the bundle is part of the cache directory. As a result, when specifying a bundle name and not a file URI, it was not possible to load a bundle without an internet connection. This especially turned out to be an issue when starting the sandbox with certain features, because the feature logic uses bundle names and not file URIs to load bundle to ConductR.

**Solution description**
This PR introduces a `—ofline` flag for the `conduct load` and `sandbox run` command to enable the offline mode. It is also possible to set the environment variable `CONDUCTR_OFFLINE_MODE=true`. Then the `—offline` flag is not necessary.

If the offline mode is enabled, then:
- The bundle is never retrieved from bintray resolver
- The `load_bundle_from_cache` and `load_bundle_configuration_from_cache` methods are resolving the last modified file for a given bundle name, e.g. visualizer. This logic is only executed if the `conduct load` command specifies a bundle name, e.g. `conduct load visualizer`, not if a file URI is specified, e.g. `conduct load /dir/my-bundle.zip`.

As a result, the sandbox and the conduct load command can be used without an internet connection in case the bundles are part of the cache directory.