Default to 1 core instance
With the sandbox run command it is possible to specify the `nr_of_instances`. When the number format was used, e.g. `-n 3` then prior to this PR 3 core and agent instances were started. Now, only 1 core instance and 3 agent instances are started.

This PR also adds an additional alias `nr-of-instances` to the existing `nr-of-containers` and `-n` arguments to specify the number of core and agent instances. Internally, we are using now the `args.nr_of_instances` arg instead of `args.nr_of_containers`.