sandbox restart command
Introducing the `sandbox restart` command. It will call first a `sandbox stop` and then a `sandbox run` based on the latest performed `sandbox run` command.

On each `sandbox run`, the args of the command is written to the file `.conductr/tmp/latest_sandbox_run_args`. This way, the restart command can pick up the run args.