Calculate loopback interface name correctly
The loopback interface name always started with `lo0`, no matter if an alias with `lo0` already exists. This resulted into false output when someone first started the sandbox with `-n 1` and afterwards with a higher number, e.g. `-n3`.

Now the loopback interface name on linux is calculated based on the corresponding IP address. If the IP address is `192.168.0.2` then the loopback name is `lo1`.

This solution works both for IP4 and IP6.