`conduct load` can now accept new `shazar` format that embeds digest at the end of the stream.
It can read this format over stdin or from a file on disk.
If encountering this file format in either scenario, `bundle.conf` name value plus the digest is used to name the bundle for conductr, thus ensuring a consistent loading experience even if the operator renames the files.
It currently uses temporary files in its implementation, but it is technically possible to work towards streaming this data directly into conductr