Catching license 503 errors
The license calls `GET /license` and `POST /license` catching now the HTTP status code 503 and then returning a `LICENSE_ENDPOINT_NOT_SUPPORTED` flag. Calling methods can then use this response to perform alternate logic.

The `conduct info` and `conduct load-license` commands have been modified accordingly.

**conduct info**
If a 503 is of `GET /license` is returned then no license info is printed. This is different to the return type None which means that no license has been uploaded to ConductR which will print the no license information

**conduct load-license**
Before posting a license we now perform a `GET /license` call to check if a 503 is returned. If so the message `conduct load-license is only supported by ConductR 2.0.3+` is printed to the console. Note that this message can be suppressed if `args.quite` is et to True. This might be helpful if the function is called from another Python function that don’t want to print any license output for ConductR versions 2.0.2-

Fixes https://github.com/typesafehub/conductr-cli/issues/340