Download license from Lightbend.com
Download license from Lightbend.com as part of `conduct load-license` command.

Cached auth token will be sent as OAuth2 Bearer token when requesting license from Lightbend.com.

If the cached auth token is not present, user will be prompted to enter the auth token along with the instruction on how to obtain the auth token. The entered token is cached for subsequent use.

If the token is expired or user has not accepted t&c, HTTP `401` and `403` is returned by the license download endpoint respectively. The response text from this endpoint will contain the instruction on how to proceed, and thus it will be presented to the user.

The command `conduct load-license` has additional switch `--license-download-url` to allow overriding the default download license endpoint URL. This option is not shown as part of help text however.