Automatically add network addressaliases
Adding the network address aliases for the user after they have entered the password for sudo priviliges. If the password has been entered 3 times incorrectly then the prompt will abort and we will display the commands to add the aliases. Also, automatically adding the aliases is only supported on linux and macOS. Otherwise, a message is displayed saying that the user need to add the aliases manually.