Hostname lookup validation pre-check
Adding additional check that the hostname lookup doesn’t take too long. This validation is necessary to check if ConductR would fail at startup given this issue with Java 8 on macOS: https://github.com/akka/akka/issues/22160

The validation is failing if the hostname is not part of the `/etc/hosts` file and if the operation system is macOS. In this case, we abort with this error:

```
$ sandbox run 2.0.2
Error: Hostname lookup took 5006 milliseconds which will result in a ConductR startup failure
Error: This is known Java 8 issue on macOS: http://stackoverflow.com/questions/39636792/jvm-takes-a-long-time-to-resolve-ip-address-for-localhost
Error: To speed up the hostname lookup add your macOS hostname to /etc/hosts
Error: Resolves your hostname on the terminal with: hostname
Error: Sample /etc/hosts file:
Error: 127.0.0.1   localhost mbpro.local
Error: ::1         localhost mbpro.local
```

We do not check for the Java version given that this error only occurs from Java 8 onwards and ConductR supports only Java 8 and onwards.

The check is only performed in sandbox JVM mode.