Handle stopping state
In ConductR 2.1, a bundle may now be stopping and thus set inActive (isStarted) to false... In this case, when a "conduct stop" command is issue, we want to wait for bundle execution entry to actually disappear before reaching the scaled state i.e. we are only waiting on isActive to become True when scaling up.