Ambigous bind protocol
Changing the validation of the endpoint bind protocol.

If the argument `--bind-protocol` is set then we use it. Otherwise, we check if we can derive the bind protocol from the acls. If all acls contain only one protocol family, e.g. `http` then we set the bind protcol to this protocol family. If multiple protocol families in the acls exist then we raise an `AmbigousBindProtocolError` exception. If no acls have been specified then we set the bind protocol to `tcp` (the default).

On the `bndl` command, if the `AmbigousBindProtocolError` is thrown, we output the following validation error:

```
bndl: argument --bind-protocol is required when acls with different protocol families are specified
endpoint: tcp
acls: tcp:[5000,5001], udp:[3000]
```