Native executable: include cacert file in the package
* Commit `.spec` files as part of the repository.
* When building in OSX, the `.spec` files for `conduct` and `sandbox` now includes cacert file as part of executable produced by the PyInstaller.
* The main class of `conduct` and `sandbox` native executable now loads the cacert from the PyInstaller package.

Include Python's cacert file as part of the packaged native executable.
This is because in OSX the OpenSSL from Macports and Homebrew doesn't fallback to CA certs stored in OSX keychain tool.

The workaround is based on this PyInstaller recipe:
https://github.com/pyinstaller/pyinstaller/wiki/Recipe-OpenSSL-Certificate

and its corresponding PR:
https://github.com/pyinstaller/pyinstaller/pull/1411/files

This workaround relies on setting SSL_CERT_FILE environment variable, and this environment variable is applicable
on Python 3.4 and 3.5: https://bugs.python.org/issue22449

Some further info on how OSX OpenSSL cert validation:
https://hynek.me/articles/apple-openssl-verification-surprises/