Check for bundle-status component
The status check component is now named `bundle-status` in the newest `sbt-conductr` version. In the `conductr-cli`, we now detecting the status check component by this name instead of searching for the substring `-status`. This change makes the detection more bulletproof, e.g. we also detect the correct status check component if the `bundle.conf` contains additional components that contain `-status` in the name.