Conduct stop also require similar fail-fast argparse option as Conduct run
Conduct stop makes use `bundle_scale.wait_for_scale()` which waits for the bundle scale to reach 0.

The same `bundle_scale.wait_for_scale()` is used for Conduct run to wait for a requested bundle scale to be achieved, which will fail fast when encountering error.

`bundle_scale.wait_for_scale` requires a number of silent argparse options are required to display bundle events and logs when error occurs.

Because of this, conduct stop requires these silent argparse options to be specified