Bundle validation
The `bndl` command is now validating the content of the `bundle.conf` in case one is created or updated given the `bndl` arguments and a possible `bundle.conf` as input.

The following bundle.conf are performed:
- Assert that all property names in the bundle.conf are known property names. This check is skipped if the bndl argument `--validation-exclude property-names` is specified
- Assert that all required bundle.conf properties are specified. This check is not performed if the format is set to `configuration` given that bundle.conf of a bundle configuration is an overlay of another bundle.conf and hence not all required properties need to be specified. This check is skipped if the bndl argument `--validation-exclude required` is specified

Also, the `bndl` implementation has changed to support all use cases described in https://docs.google.com/document/d/11aEKF8SUv-7f7HrOypAG7CICJLXIhqA-6M14kdzt_dM/edit. In particular the following use cases are now supported as well:
- Create a bundle configuration from scratch with bndl arguments (no input)
- Modiying an existing bundle configuration that does not contain a bundle.conf file
- Creating a bundle or bundle configuration given an input directory that does not contain a bundle.conf, e.g. an application directory

Note that when creating a bundle configuration, the format need to be explicitly set to `--format configuration`. The format auto-detection of the bndl command cannot distinguish between a `bundle` and a `configuration`. Because the default use case of the `bndl` command is to create or modify a bundle, the format auto-detection will fallback to `bundle`.