Native executable: use certifi to provide cacert
This is because when packaging as native executable, the `ssl.get_default_verify_paths().cafile` may return `None` depending on where the platform the packaging is performed on.

On Ubuntu 14 and Ubuntu 16, the `ssl.get_default_verify_paths().cafile` returns `None`, which means there are no cacert package in the Linux CLI executable. When this CLI executable is ran on other flavour of Linux (i.e. CoreOS), the HTTPS url access will fail due to SSL cert validation error.

By importing `certifi`, the cacert will be included as part of the `certifi`, and the `certifi`'s cacert is used when the CLI is ran as native executable.

The pip3 install of the CLI will retain its default behaviour, i.e. rely on the cacert supplied by default on the machine's Python 3.