Allow empty services and acls arrays
The bundle validation failed when an endpoint contained an empty array in the `services` or `acls` fields. However, empty fields are valid here, especially considering that a bundle configuration can declare those values to override existing `services` or `acls` fields with an empty array.

This change now allows empty `services` and `acls` fields.